change: OrganizationLogo as component
