+++ src/components/UniversiHeader/styles.less
@@ -27,7 +27,8 @@
         border-top-left-radius:    0;
         border-bottom-left-radius: 0;
 
-
+        font-size: 1.75rem;
+        color: @font-color-v2;
     }
 
     .logo-container + .left-items {
